pyproject.toml
[project]
name = "image-browser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "autopep8>=2.3.2",
    "flake8>=7.3.0",
    "flask>=3.1.2",
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "pillow>=12.0.0",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
]
[tool.pycodestyle]
max-line-length = 120
ignore = ["E251", "E221"]

[tool.isort]
include_trailing_comma = false
use_parentheses = false
line_length = 120

[tool.coverage.run]
omit = ["*/test/*"]

